+++ extensions/timers/src/types.ts
@@ -28,6 +28,7 @@ export interface CustomTimer {
 export interface Preferences {
   selectedSound: string;
   ringContinuously: boolean;
+  copyOnSwStop: boolean;
 }
 
 export interface CTInlineArgs {